auth2/keys/serializers.py
from rest_framework import serializers
from .models import APIKey

class APIKeySerializer(serializers.ModelSerializer):
    class Meta:
        model = APIKey
        fields = ['id', 'name', 'key', 'is_active', 'created_at']
        read_only_fields = ['key', 'created_at']
